Make the schema files JSON
This undoes a change made in PR #217.

Although JSON is a disasterous syntax for human-maintained files, the fact
remains that JSON-schema is a thing, whilst YAML-schema (although accepted by
swagger) is not. Using YAML negates most of the advantages of using the
JSON-schema format (mostly, the fact that we can feed those files into a number
of tools), and we weren't really using the increased power of YAML.

Possibly, we should consider generating the JSON from YAML inputs, but either
way, the event schemas need to refer to the JSON files, and for now it's easier
just to stick to JSON.