Don't line wrap the API descriptions
They are in RST format which is whitespace sensitive and threfore can't be line wrapped without breaking the syntax.

There is more wrapping in other places but some of it is into tables where it does need to be fixed width in order for the RST table to parse correctly.