Name newly-empty rooms based on last user to leave
This relieves clients from tracking the total number of users who have
ever been in a room and the name of the earliest-sorted user, and
ensures that the display name change resulting from the transition from
a 1:1 room to an empty room is intuitive. Wording is also tweaked to
ensure that the above holds if the last user in a room was banned.