Ensure the event examples and matrix.org assets are tested
Otherwise we go nearly a week without realizing the build is failing due to bad schemas/examples. This also helps us ensure a PR is up to par.

This commit is expected to cause a build failure at this time. A future commit will actually fix the project and address concerns raised by the testing.