Add permalink routing through ?via parameters on matrix.to URIs
Spec for [MSC1704](https://github.com/matrix-org/matrix-doc/pull/1704)

Reference implementations:
* Original: https://github.com/matrix-org/matrix-react-sdk/pull/2250
* Modern recommendations: https://github.com/matrix-org/matrix-react-sdk/blob/2ca281f6b7b735bc96945370584c5cb1b5f7e1f1/src/matrix-to.js#L29-L70

The only deviation from the original MSC is the recommendation for which servers to pick. The original MSC failed to consider server ACLs and IP addresses correctly, and during implementation it was realized that both of these cases should be handled. The core principles of the original MSC are left unaltered.