Specify the capabilities action
This diverges slightly from the MSC to match reality in that the MSC wants the widget to get all the permissions it asked for or be rejected, however in practice the client can deny some/all of the permissions without notifying the widget and the widget is expected to continue normally.

This deviation is generally fine as most capabilities are in the client->widget direction anyhow.