Specify versioning for the widget API
This has significant deviation from the MSC and practical implementation due to lack of validation. The MSC specifies the request/response schema as described in this commit, though the associated behaviour and timings are different.

In practice, no widget has implemented `supported_api_versions` and no client has ever requested it from a widget. This is expected to change with the specification being present. Additionally, widgets are technically able to send an `api_version` request to get the *current* version, though this is implied by the `supported_api_versions` response and thus not considered.

Tying the version to the spec release is something introduced by this commit, as are the restrictions about what can and cannot be sent by either side. This commit also introduces timing constraints that allow widgets/clients to check the other's versions at any time, even during session initialization, due to the potential need for the initialization sequence to change.

Further, this lumps `0.0.1` and `0.0.2` into the same version (which is expected to release as `0.1.0`) despite minor differences between the two. These differences are years old and effectively not important in the eyes of the spec.

It is arguable that this particular section should be an MSC on top of the widgets MSC rather than landing like this.