Drop `_room` suffix
Matthew wants us to drop the `_room` suffix. I worry that this is somewhat
confusing (it's not obvious to the casual observer what it is that we are
peeking), and risks a lot of confusion in the future when we find something
else to peek, but he seems to feel strongly.