Serve Inter font locally rather than pulling from google fonts
Due to a small change in our fork of docsy, we can now override the global font URL
from a google fonts one to a local one. We do so in our SCSS variables file, as well
as add the relevant font CSS and woff2 files, and a script that generated them.