Update unstable endpoints from i.e xyz.amorgan/knock -> xyz.amorgan.knock
I'm doing this for two reasons:

1) If I ever wanted to create another knock-related MSC, it would be nice to be able
to give it a different namespace for endpoints, such as xyz.amorgan.knock2/knock. If
we were only using xyz.amorgan as the namespace for endpoints, that restricts that
namespace to only work for one MSC which attempts to use /knock as part of an endpoint.

2) I accidentally made the implementation use xyz.amorgan.knock/knock :)