Fix inconsistency of blurhashes MUST vs SHOULD be omitted
I justify this being a MUST as otherwise naive clients see and present
a blurhash, but never show the actual image as it wouldn't be available.

This may believe one to believe that their media repository is broken.
However, in fact the event itself just has a blurhash without a
corresponding image.