Use changelogs/release.yaml for the version number source
Previously both the top banner of the site and the top paragraph on
the changelog page would pull the current version number of the spec
from config.toml.

This commit changes the source of truth to changelogs/release.yaml,
which is already used to determine the current version of the spec
on other parts of the changelogs page.

This also means one less thing to change when doing a release.