Remove gilesbot config
gilesbot is a tool written and run by @Cadair that would insert little "click here to preview the spec content" entries in the CI checks on a PR. Nowadays, however, we're started using Github Actions and netlify, with the netlify preview link being added to the PR description by a Github Action.

As such, gilesbot isn't necessary for our purposes anymore. But it was really useful while it lasted!

This commit removes the config for it, and as that was the only thing in `pyproject.toml`, the whole file as well.