Allowing widgets to read account data
This proposal gives a way for widgets to read account data.

Some data is stored in the account data that is useful for widgets, such as
MSC2545, which stores a list of image packs in account data.

If this proposal is implemented, a widget can read all global packs defined in MSC2545 by
additionally using MSC2876 which is already implemented.