Remove the superfluous UPDATE command
We previously needed this command because room
data HAD to exist inside an `op`, so if a room
was updated without changing position then it
needed a command to say "don't change its position"
which _is_ the UPDATE command. Now that room data
sits outside the `ops`, we no longer need the
UPDATE command at all.