Add a stable flag to MSC3916
In conversation with several client and homeserver developers it has
come up that the current schedule for rolling out authenticated media is
hard for them to follow if they have to support all of 1.11 at once.

I think the value of authenticated media is high enough to warrant a
stable flag to encourage a faster rollout of authenticated media. This
allows servers to support authenticated media without supporting all of
1.11 already and the additional burden on client developers is minimal.

This flag is already in use by serveral implementations:

- https://github.com/neilalexander/harmony/commit/51b5c9803344fa1f33832cd6ccacd37db383ec84
- https://github.com/famedly/matrix-dart-sdk/commit/2dce08bab1006abe27d1931f4ba74a7fb693873c

Signed-off-by: Nicolas Werner <nicolas.werner@hotmail.de>