Reword 'identity was changed' to 'identity has been reset'
Based on real-world feedback from Element clients, 'identity was changed', and
especially 'identity appears to have been changed' were very confusing.

Resetting an identity is something that the user can do themselves, so it is
much clearer what we mean. They can ask the other user "did you reset your
identity?" and it is easier for them to answer.