Clarify the meaning of a read_events action without a state_key (#4288)
Currently MSC2762 is ambiguous as to whether the absence of a state_key in a read_events action means that both state events and message-like events should be returned, or just message-like events. We should clarify this behavior, and I for one find it difficult to imagine a real use case that would benefit from the ability to read both state events and message-like events of a single event type as part of the same request, so I propose we limit the response to just message-like events in this case. This aligns the behavior with what Element Web did prior to the introduction of the update_state action.