Update QR code format under "type" 0x03
The byte immediately following the `MATRIX` prefix is repurposed as a type which gives a more sensible way to namespace in future.

This also means that we can call the second byte the "intent" very clearly rather than having to call it "mode" to match the existing cross verification QR spec.